refactor: rename safety module to process documents
Merge safety.types.ts and safety.service.ts into their process document
counterparts. Rename store (useJsaSwpStore), views (JsaListView,
SwpListView), and move safety-wizard components under
process-documents/. Update router paths, navbar, and all imports.
