+++ src/components/job/JobSafetyTab.vue
@@ -132,7 +132,7 @@
     v-if="wizardDocument"
     :is-open="isWizardOpen"
     :document-id="wizardDocument.id"
-    :document-type="wizardDocument.document_type"
+    :document-type="wizardDocument.document_type as SafetyDocumentType"
     @close="closeWizard"
     @saved="handleWizardSaved"
   />
@@ -161,9 +161,9 @@ import {
   DialogDescription,
   DialogFooter,
 } from '@/components/ui/dialog'
-import SafetyWizardModal from '@/components/safety/SafetyWizardModal.vue'
-import { useSafetyDocumentsStore } from '@/stores/safetyDocuments'
-import type { SafetyDocumentList } from '@/types/safety.types'
+import SafetyWizardModal from '@/components/process-documents/safety-wizard/SafetyWizardModal.vue'
+import { useJsaSwpStore } from '@/stores/jsaSwpDocuments'
+import type { ProcessDocumentListItem, SafetyDocumentType } from '@/types/processDocument.types'
 import { formatDate } from '@/utils/string-formatting'
 
 interface Props {
@@ -174,18 +174,18 @@ interface Props {
 const props = defineProps<Props>()
 
 const emit = defineEmits<{
-  'jsa-generated': [doc: SafetyDocumentList]
+  'jsa-generated': [doc: ProcessDocumentListItem]
   'jsa-deleted': [docId: string]
 }>()
 
 // Store
-const store = useSafetyDocumentsStore()
+const store = useJsaSwpStore()
 
 // Local state
 const isDeleteDialogOpen = ref(false)
-const documentToDelete = ref<SafetyDocumentList | null>(null)
+const documentToDelete = ref<ProcessDocumentListItem | null>(null)
 const isWizardOpen = ref(false)
-const wizardDocument = ref<SafetyDocumentList | null>(null)
+const wizardDocument = ref<ProcessDocumentListItem | null>(null)
 
 // Computed
 const documents = computed(() => store.getDocumentsByJobId(props.jobId))
@@ -239,13 +239,13 @@ async function handleGenerateJSA() {
   }
 }
 
-function openInGoogleDocs(doc: SafetyDocumentList) {
+function openInGoogleDocs(doc: ProcessDocumentListItem) {
   if (doc.google_doc_url) {
     window.open(doc.google_doc_url, '_blank')
   }
 }
 
-function openWizard(doc: SafetyDocumentList) {
+function openWizard(doc: ProcessDocumentListItem) {
   wizardDocument.value = doc
   isWizardOpen.value = true
 }
@@ -261,7 +261,7 @@ function handleWizardSaved() {
   toast.success('Document saved successfully!')
 }
 
-function confirmDelete(doc: SafetyDocumentList) {
+function confirmDelete(doc: ProcessDocumentListItem) {
   documentToDelete.value = doc
   isDeleteDialogOpen.value = true
 }
