+++ src/router/index.ts
@@ -288,15 +288,15 @@ const router = createRouter({
       meta: { requiresAuth: true, title: 'Process Document - Jobs Manager' },
     },
     {
-      path: '/safety/jsa',
-      name: 'safety-jsa',
-      component: () => import('@/views/SafetyJsaView.vue'),
+      path: '/process-documents/jsa',
+      name: 'jsa-list',
+      component: () => import('@/views/JsaListView.vue'),
       meta: { requiresAuth: true, title: 'Job Safety Analyses - Jobs Manager' },
     },
     {
-      path: '/safety/swp',
-      name: 'safety-swp',
-      component: () => import('@/views/SafetySwpView.vue'),
+      path: '/process-documents/swp',
+      name: 'swp-list',
+      component: () => import('@/views/SwpListView.vue'),
       meta: { requiresAuth: true, title: 'Safe Work Procedures - Jobs Manager' },
     },
   ],
